requirements.txt
requests
beautifulsoup4
lxml
python-telegram-bot==13.15

